[#229] refactor: 폴더 구조 변경 및 DebateDetail 부분 코드 리팩토링
