+++ src/app/debate-list/[debateId]/components/Comments/CommentCard.tsx
@@ -3,8 +3,8 @@ import { usePathname } from 'next/navigation';
 import ReportModal from '@/components/Common/ReportModal';
 import { deleteComment, updateComment } from '@/service/debate/comment';
 import { Box, useDisclosure } from '@chakra-ui/react';
-import EditCommentForm from './EditCommentForm';
-import CommentDisplay from './CommnetDisplay';
+import EditCommentForm from './CreateComment/EditCommentForm';
+import CommentDisplay from './CreateComment/CommnetDisplay';
 
 export interface DebateCommentProps {
   userImg: string;
