+++ src/app/debate-list/[debateId]/components/DebateDetail/DebateDetail.tsx
@@ -1,10 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { Amendment } from '@/types/common/Amendment';
 import transTitleTagtoAmendment from '@/lib/debate/transTitleAndTagChanged';
-import DebateInformation from './DebateInformation';
-import DebateSlider from './Carousel/DebateSlider';
-import MiddleTitle from './MiddleTitle';
-import { Debate } from '../page.server';
+import DebateInformation from './DebateInfo/DebateInformation';
+import DebateSlider from './DebateSlider/DebateSlider';
+import { Debate } from '../../page.server';
 
 interface DeabteDetailProps {
   debateData: Debate | null;
@@ -48,24 +47,17 @@ const DebateDetail: React.FC<DeabteDetailProps> = ({ debateData }) => {
   }, [debateData]);
 
   return (
-    <div className="flex-col ">
+    <div
+      className={`flex flex-col p-10 pb-20 rounded-lg mb-10 text-white border-2 border-primary-dark-500/20 ${debateData?.status === 'CLOSED' ? 'opacity-60' : ''}`}
+    >
       {debateData && (
-        <div
-          className={`flex flex-col px-10 pt-8 pb-20 rounded-lg text-white border-2 border-primary-dark-500/20 mb-10 ${debateData.status === 'CLOSED' ? 'opacity-60' : ''}`}
-        >
+        <>
           <DebateInformation
             contributeData={debateData.contribute}
             debateEndAt={debateData.endAt}
           />
-          <div className="mt-10 my-5">
-            <MiddleTitle title="수정 요청 사항" color="white" />
-            <div className="tiptap flex justify-between gap-5 text-primary-dark-500 ">
-              <p className="font-bold w-1/2 text-center mb-4 ">수정 전</p>
-              <p className="font-bold w-1/2 text-center mb-4">수정 후</p>
-            </div>
-          </div>
           {totalAmendments && <DebateSlider amendments={totalAmendments} />}
-        </div>
+        </>
       )}
     </div>
   );
