+++ src/app/debate-list/page.tsx
@@ -6,7 +6,7 @@ import Link from 'next/link';
 import { Pagination } from '@mui/material';
 import { Center } from '@chakra-ui/react';
 import Wrapper from '@/components/Common/Wrapper';
-import PageTitle from '@/components/Common/PageTitle';
+import PageTitle from '@/components/Common/Title/PageTitle';
 import StateTab from '@/components/Common/StateTab';
 import ChakraSelect from '@/components/Common/ChakraSelect';
 import DebateListCard from '@/app/debate-list/components/DebateListCard';
